libdispatch-1271.40.12
Imported from libdispatch-1271.40.12.tar.gz

+++ dispatch/queue.h
@@ -681,7 +681,7 @@ typedef long dispatch_queue_priority_t;
 API_AVAILABLE(macos(10.6), ios(4.0))
 DISPATCH_EXPORT DISPATCH_CONST DISPATCH_WARN_RESULT DISPATCH_NOTHROW
 dispatch_queue_global_t
-dispatch_get_global_queue(long identifier, unsigned long flags);
+dispatch_get_global_queue(intptr_t identifier, uintptr_t flags);
 
 /*!
  * @typedef dispatch_queue_attr_t
